requirements-dev.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.4.2
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.3
django-ajax-datatable==4.1.4
django-background-tasks==1.2.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.0.1
django-compat==1.0.15
django-cors-headers==3.7.0
django-crispy-forms==1.9.1
django-datatables-view==1.19.1
django-excel==0.0.10
django-excel-to-model==0.1.6
django-import-export==2.2.0
django-jsonfield==1.4.1
django-jsonify==0.3.0
django-multiselectfield==0.1.12
django-npm==1.0.0
django-phone-field==1.8.1
django-tinymce==3.1.0
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
httpie==2.4.0
idna==2.10
ipython==7.12.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.16.0
jsmin==2.2.2
jupyter-core==4.7.1
lazy-object-proxy==1.4.3
lml==0.1.0
MarkupPy==1.14
numpy==1.20.1
odfpy==1.4.1
openpyxl==3.0.6
pandas==1.2.2
parso==0.8.1
pickleshare==0.7.5
pinyin==0.4.0
prompt-toolkit==3.0.16
python-decouple==3.4
pytz==2021.1
soupsieve==2.0.1
sqlparse==0.4.1
static3==0.7.0
tablib==2.0.0
texttable==1.6.2
traitlets==4.3.3
urllib3==1.26.3
wcwidth==0.1.8
webencodings==0.5.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
